Added install confirmation prompt and uninstall script

+++ install.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 dir=$(pwd -P);
 cmd=$(basename $dir);
-ln -s $dir/bin/$cmd /usr/local/bin/$cmd;
+read -p "Create symbolic link to $dir/bin/$cmd in /usr/local/bin/$cmd? " -n 1 -r REPLY;
+echo;
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+	ln -s $dir/bin/$cmd /usr/local/bin/$cmd;
+fi
